+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - Remove profiling timeout logic and disable profiling on API 21 ([#3478](https://github.com/getsentry/sentry-java/pull/3478))
+- Properly reset metric flush flag on metric emission ([#3493](https://github.com/getsentry/sentry-java/pull/3493))
 
 ## 7.10.0